Rename 'gulpfile.js' of prototype-kit to 'defineGulpTasks.js' since it is nolonger a standalone gulp file.
