+++ defineGulpTasks.js
